Refactor: tidying project structure, removing duplications
