+++ plant-care-assistant/app/login/page.tsx
@@ -1,4 +1,4 @@
-import LoginForm from '@/components/Forms/LoginForm';
+import LoginForm from "@components/forms/LoginForm";
 
 export default function LoginPage() {
   return (
