--- plant-care-assistant/app/myfeed/page.tsx
@@ -1,11 +1,0 @@
-"use client";
-
-export default function MyFeed() {
-  return (
-    <div className="flex items-center justify-center h-screen">
-      <p className="text-gray-600 text-xl font-medium">
-        Here I will be able to see all my posts.
-      </p>
-    </div>
-  );
-}
