+++ plant-care-assistant/app/register/page.tsx
@@ -1,4 +1,4 @@
-import RegisterForm from '@/components/Forms/RegisterForm';
+import RegisterForm from "@components/forms/RegisterForm";
 
 export default function RegistrationPage() {
   return (
