--- plant-care-assistant/components/LoginView/Login.tsx
@@ -1,83 +1,0 @@
-"use client";
-
-import React, { useState } from "react";
-import { useAuth } from "../../providers/AuthProvider";
-import { useRouter } from "next/navigation";
-
-const Login = () => {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [error, setError] = useState("");
-  const [isLoading, setIsLoading] = useState(false);
-  const { signIn } = useAuth();
-  const router = useRouter();
-
-  const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-    setError("");
-    setIsLoading(true);
-
-    try {
-      await signIn(email, password);
-      router.push("/dashboard"); // Redirect to dashboard after login
-    } catch (err) {
-      console.error("Login error:", err);
-      setError(err instanceof Error ? err.message : "Failed to sign in");
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  return (
-    <div className="bg-darkest p-8 rounded-lg shadow-lg w-full h-full flex flex-col justify-center items-center">
-      <h2 className="text-2xl font-bold text-plant mb-6 text-center">Login</h2>
-      
-      {error && (
-        <div className="mb-4 p-3 bg-red-100 text-red-700 rounded-md">
-          {error}
-        </div>
-      )}
-      
-      <form
-        onSubmit={handleSubmit}
-        className="flex flex-col gap-4 w-full max-w-md"
-      >
-        <div className="flex flex-col">
-          <label htmlFor="email" className="text-plant mb-2">
-            Email:
-          </label>
-          <input
-            type="email"
-            name="email"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            className="p-2 rounded-md border border-gray-300 focus:outline-none focus:ring-2 focus:ring-verdigris-light"
-            required
-          />
-        </div>
-        <div className="flex flex-col">
-          <label htmlFor="password" className="text-plant mb-2">
-            Password:
-          </label>
-          <input
-            type="password"
-            name="password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            className="p-2 rounded-md border border-gray-300 focus:outline-none focus:ring-2 focus:ring-verdigris-light"
-            required
-          />
-        </div>
-        <button
-          type="submit"
-          disabled={isLoading}
-          className="bg-verdigris-light text-plant py-2 px-4 rounded-md hover:bg-verdigris-dark transition duration-300"
-        >
-          {isLoading ? "Signing in..." : "Login"}
-        </button>
-      </form>
-    </div>
-  );
-};
-
-export default Login;
