+++ plant-care-assistant/components/common/Icon.tsx
