+++ plant-care-assistant/components/common/Logo.tsx
