+++ plant-care-assistant/components/features/calendar/styles/calendar.css
