+++ plant-care-assistant/components/layout/Header.tsx
@@ -1,13 +1,13 @@
-'use client'
+"use client";
 
-import Icon from '@/components/icon'
-import Link from 'next/link'
-import { useRef, useState } from 'react'
-import DashboardModal from '@/components/ui/DashboardModal'
+import Icon from "@components/common/Icon";
+import Link from "next/link";
+import { useRef, useState } from "react";
+import DashboardModal from "@components/ui/DashboardModal";
 
 export default function Header() {
-  const dashboardRef = useRef<HTMLAnchorElement>(null)
-  const [isModalOpen, setIsModalOpen] = useState(false)
+  const dashboardRef = useRef<HTMLAnchorElement>(null);
+  const [isModalOpen, setIsModalOpen] = useState(false);
 
   return (
     <header className="fixed top-0 left-0 w-full bg-[#0A1A24] px-4 py-3 flex justify-between items-center h-16 z-50">
@@ -29,5 +29,5 @@ export default function Header() {
         anchorRef={dashboardRef}
       />
     </header>
-  )
+  );
 }
